Change WebpImage methods to use size_t

+++ include/exiv2/webpimage.hpp
@@ -88,8 +88,7 @@ namespace Exiv2 {
         void doWriteMetadata(BasicIo& outIo);
         //! @name NOT Implemented
         //@{
-        long getHeaderOffset(byte *data, long data_size,
-                             byte *header, long header_size);
+        long getHeaderOffset(byte* data, size_t data_size, byte* header, size_t header_size);
         bool equalsWebPTag(Exiv2::DataBuf& buf ,const char* str);
         void debugPrintHex(byte *data, long size);
         void decodeChunks(uint64_t filesize);
